+++ noodles-cram/src/async/reader/record.rs
@@ -128,8 +128,10 @@ where
         };
 
         if reference_id != bam::record::reference_sequence_id::UNMAPPED {
-            let reference_sequence_id = bam::record::ReferenceSequenceId::try_from(reference_id)
+            let reference_sequence_id = usize::try_from(reference_id)
+                .map(bam::record::ReferenceSequenceId::from)
                 .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+
             builder = builder.set_reference_sequence_id(reference_sequence_id);
         }
 
@@ -353,7 +355,8 @@ where
             if id == bam::record::reference_sequence_id::UNMAPPED {
                 Ok(None)
             } else {
-                bam::record::ReferenceSequenceId::try_from(id)
+                usize::try_from(id)
+                    .map(bam::record::ReferenceSequenceId::from)
                     .map(Some)
                     .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
             }