+++ noodles-cram/src/crai/async/reader.rs
@@ -114,7 +114,7 @@ mod tests {
         let actual = read_index(&mut reader).await?;
 
         let expected = vec![Record::new(
-            bam::record::ReferenceSequenceId::try_from(0).map(Some)?,
+            Some(bam::record::ReferenceSequenceId::from(0)),
             10946,
             6765,
             17711,